+++ Fritz.StaticBlog/ActionBuild.cs
@@ -1,14 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Text;
 using System.Text.Json;
 using CommandLine;
 using Markdig;
-using Markdig.Extensions.Yaml;
 using Markdig.Prism;
-using Markdig.Syntax;
 
 namespace Fritz.StaticBlog;
