+++ Fritz.StaticBlog/Data/LastBuild.cs
@@ -1,0 +1,8 @@
+namespace Fritz.StaticBlog.Data;
+
+public class LastBuild
+{
+
+	public DateTime Timestamp { get; set; }
+
+}
