--- Fritz.StaticBlog/LastBuild.cs
@@ -1,13 +1,0 @@
-using System;
-
-namespace Fritz.StaticBlog
-{
-
-	public class LastBuild
-	{
-
-		public DateTime Timestamp { get; set; }
-
-	}
-
-}
