+++ Test.StaticBlog/GivenLastBuildFile/WhenBuilding.cs
@@ -1,29 +1,21 @@
-using System;
-using System.IO;
-using Xunit;
+namespace Test.StaticBlog.GivenLastBuildFile;
 
-namespace Test.StaticBlog.GivenLastBuildFile
+public class WhenBuilding : BaseFixture
 {
 
-	public class WhenBuilding : BaseFixture
+	[Fact]
+	public void ShouldReadLastBuildDate()
 	{
 
-		[Fact]
-		public void ShouldReadLastBuildDate()
-		{
+		// arrange
+		Initialize();
 
-			// arrange
-			Initialize();
+		// act
+		_sut.Validate();
 
-			// act
-			_sut.Validate();
-
-			// assert
-			Assert.Equal(_LastBuildDate, _sut._LastBuild?.Timestamp);
-
-		}
+		// assert
+		Assert.Equal(_LastBuildDate, _sut._LastBuild?.Timestamp);
 
 	}
-
 
 }
