+++ Test.StaticBlog/GivenSiteWithoutLastBuild/WhenBuilding.cs
@@ -1,60 +1,54 @@
-using System;
-using System.IO;
 using Fritz.StaticBlog;
-using Xunit;
 
-namespace Test.StaticBlog.GivenSiteWithoutLastBuild 
+namespace Test.StaticBlog.GivenSiteWithoutLastBuild;
+
+public class WhenBuilding : TestSiteBaseFixture, IDisposable
 {
-		
-	public class WhenBuilding : TestSiteBaseFixture, IDisposable
+	private readonly ActionBuild _sut;
+
+	public WhenBuilding()
 	{
-		private readonly ActionBuild _sut;
-
-		public WhenBuilding()
-		{
-				
-			base.Initialize();
-
-			_sut = new ActionBuild
-			{
-				Force = false,
-				OutputPath = TargetFolderName,
-				ThisDirectory = WorkingDirectory.FullName,
-				LastBuildFilename = $".lastbuild.{Guid.NewGuid()}.json",
-				Config = new Config
-				{
-					Theme = "kliptok",
-					Title = "The Unit Test Website"
-				}
-			};
-
-		}
- 
-		public void Dispose()
-		{
 			
-			File.Delete(Path.Combine(WorkingDirectory.FullName, _sut.LastBuildFilename));
+		base.Initialize();
 
-		}
+		_sut = new ActionBuild
+		{
+			Force = false,
+			OutputPath = TargetFolderName,
+			ThisDirectory = WorkingDirectory.FullName,
+			LastBuildFilename = $".lastbuild.{Guid.NewGuid()}.json",
+			Config = new Config
+			{
+				Theme = "kliptok",
+				Title = "The Unit Test Website"
+			}
+		};
 
-		[Fact]
-		public void ThenTheLastBuildFileShouldBeGenerated() {
+	}
+ 
+	public void Dispose()
+	{
+		
+		File.Delete(Path.Combine(WorkingDirectory.FullName, _sut.LastBuildFilename));
 
-			_sut.Execute();
+	}
 
-			var lastBuildFile = Path.Combine(WorkingDirectory.FullName, _sut.LastBuildFilename);
-			Assert.True(System.IO.File.Exists(lastBuildFile));
+	[Fact]
+	public void ThenTheLastBuildFileShouldBeGenerated() {
 
-		}
+		_sut.Execute();
 
-		[Fact]
-		public void ThenTheLastBuildShouldBeMinDate() {
+		var lastBuildFile = Path.Combine(WorkingDirectory.FullName, _sut.LastBuildFilename);
+		Assert.True(System.IO.File.Exists(lastBuildFile));
 
-			_sut.Validate();
+	}
 
-			Assert.Equal(DateTime.MinValue, _sut._LastBuild.Timestamp);
+	[Fact]
+	public void ThenTheLastBuildShouldBeMinDate() {
 
-		}
+		_sut.Validate();
+
+		Assert.Equal(DateTime.MinValue, _sut._LastBuild.Timestamp);
 
 	}
