+++ Test.StaticBlog/GivenValidActionBuild/BaseFixture.cs
@@ -1,54 +1,50 @@
-using System;
-using System.IO;
 using System.Reflection;
-using System.Threading;
 using Fritz.StaticBlog;
 
-namespace Test.StaticBlog.GivenValidActionBuild
+namespace Test.StaticBlog.GivenValidActionBuild;
+
+public abstract class BaseFixture
 {
-	public abstract class BaseFixture {
 
-		internal ActionBuild _sut;
-		private static ReaderWriterLockSlim folderLock = new();
+	internal ActionBuild _sut;
+	private static ReaderWriterLockSlim folderLock = new();
 
 
-		public BaseFixture()
+	public BaseFixture()
+	{
+
+		var workingDirectory = Assembly.GetAssembly(GetType()).Location.Contains(@"\.vs\") ?
+				@"..\..\..\..\..\..\..\..\..\TestSite" :
+				"../../../../TestSite";
+
+
+		string targetFolderName = GetType().Name.ToLowerInvariant();
+		_sut = new ActionBuild
 		{
-
-			var workingDirectory = Assembly.GetAssembly(GetType()).Location.Contains(@"\.vs\") ?
-					@"..\..\..\..\..\..\..\..\..\TestSite" :
-					"../../../../TestSite";
-
-
-			string targetFolderName = GetType().Name.ToLowerInvariant();
-			_sut = new ActionBuild
+			Force = false,
+			OutputPath = targetFolderName,
+			ThisDirectory = workingDirectory,
+			Config = new Config
 			{
-				Force = false,
-				OutputPath = targetFolderName,
-				ThisDirectory = workingDirectory,
-				Config = new Config
-				{
-					Theme = "kliptok",
-					Title = "The Unit Test Website"
-				}
-			};
-
-			OutputFolder = new DirectoryInfo(Path.Combine(_sut.ThisDirectory, targetFolderName));
-			if (!OutputFolder.Exists)
-			{
-				folderLock.EnterWriteLock();
-				OutputFolder.Create();
-				folderLock.ExitWriteLock();
+				Theme = "kliptok",
+				Title = "The Unit Test Website"
 			}
-			OutputPostsFolder = new DirectoryInfo(Path.Combine(OutputFolder.FullName, "posts"));
-			OutputRssFile = new FileInfo(Path.Combine(OutputFolder.FullName, "rss.xml"));	
+		};
 
+		OutputFolder = new DirectoryInfo(Path.Combine(_sut.ThisDirectory, targetFolderName));
+		if (!OutputFolder.Exists)
+		{
+			folderLock.EnterWriteLock();
+			OutputFolder.Create();
+			folderLock.ExitWriteLock();
 		}
-
-		public DirectoryInfo OutputFolder { get; private set; }
-		public DirectoryInfo OutputPostsFolder { get; private set; }
-		public FileInfo OutputRssFile { get; private set; }
+		OutputPostsFolder = new DirectoryInfo(Path.Combine(OutputFolder.FullName, "posts"));
+		OutputRssFile = new FileInfo(Path.Combine(OutputFolder.FullName, "rss.xml"));
 
 	}
+
+	public DirectoryInfo OutputFolder { get; private set; }
+	public DirectoryInfo OutputPostsFolder { get; private set; }
+	public FileInfo OutputRssFile { get; private set; }
 
 }
