+++ Test.StaticBlog/Test.StaticBlog.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-
     <IsPackable>false</IsPackable>
+		<ImplicitUsings>true</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
@@ -18,6 +18,12 @@
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
   </ItemGroup>
+
+	<ItemGroup>
+		<Using Include="Fritz.StaticBlog.Data" />
+		<Using Include="Xunit" />
+		<Using Include="Xunit.Abstractions" />
+	</ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Fritz.StaticBlog\Fritz.StaticBlog.csproj" />
